+++ bulma/index.js
@@ -3,6 +3,7 @@ import EnsoForm from '../src/bulma/EnsoForm.vue';
 import FormField from '../src/bulma/parts/FormField.vue';
 import FormFieldGroup from '../src/bulma/parts/FormFieldGroup.vue';
 import Action from '../src/bulma/parts/Action.vue';
+import CheckboxField from '../src/bulma/fields/CheckboxField.vue';
 import DateField from '../src/bulma/fields/DateField.vue';
 import InputField from '../src/bulma/fields/InputField.vue';
 import MoneyField from '../src/bulma/fields/MoneyField.vue';
@@ -12,6 +13,6 @@ import TextareaField from '../src/bulma/fields/TextareaField.vue';
 import TimeField from '../src/bulma/fields/TimeField.vue';
 
 export {
-    VueForm, EnsoForm, FormField, FormFieldGroup, Action, DateField, InputField,
+    VueForm, EnsoForm, FormField, FormFieldGroup, Action, CheckboxField, DateField, InputField,
     MoneyField, SelectField, SwitchField, TextareaField, TimeField,
 };